Update compensation property salary to be type int.
It seems more reasonable for a salary to be an integer value than for it to have decimal places.

+++ code-challenge/Models/Compensation.cs
@@ -9,7 +9,7 @@ namespace challenge.Models
     {
         public String CompensationId { get; set; }
         public Employee Employee { get; set; }
-        public Double Salary { get; set; }
+        public int Salary { get; set; }
         public DateTime EffectiveDate { get; set; }
     }
 }
